move exception classes to own package
