+++ src/main/java/task_2/OlivandersShop.java
@@ -8,6 +8,9 @@ import java.util.Map;
 import java.util.Optional;
 
 import lombok.extern.slf4j.Slf4j;
+import task_2.exceptions.NotEnoughMaterialException;
+import task_2.exceptions.NotEnoughWoodException;
+import task_2.exceptions.OrdersListIsEmptyException;
 
 @Slf4j
 public class OlivandersShop {
