+++ src/main/java/task_2/exceptions/OrdersListIsEmptyException.java
@@ -1,4 +1,4 @@
-package task_2;
+package task_2.exceptions;
 
 public class OrdersListIsEmptyException extends RuntimeException {
